+++ order-service/src/main/java/com/lotfresh/orderservice/dto/ProductRequest.java
@@ -1,5 +1,8 @@
 package com.lotfresh.orderservice.dto;
 
+import com.lotfresh.orderservice.domain.Order;
+import com.lotfresh.orderservice.domain.ProductOrder;
+import com.lotfresh.orderservice.domain.ProductOrderId;
 import lombok.AccessLevel;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -14,4 +17,19 @@ public class ProductRequest {
     public static ProductRequest forTest(Long productId, Long productPrice, Long productQuantity){
         return new ProductRequest(productId, productPrice, productQuantity);
     }
+
+    public ProductOrder toEntity(Order order) {
+        return ProductOrder.builder()
+                .id(buildProductOrderId(productId))
+                .order(order)
+                .price(productPrice)
+                .quantity(productQuantity)
+                .build();
+    }
+    private ProductOrderId buildProductOrderId(Long productId) {
+        return ProductOrderId.builder()
+                .productId(productId)
+                .build();
+    }
+
 }
