+++ order-service/src/main/resources/application.yml
@@ -6,7 +6,7 @@ spring:
     name: order-service
   jpa:
     hibernate:
-      ddl-auto: create
+      ddl-auto: update
     properties:
       hibernate:
         format_sql: true
